+++ src/trace/event.rs
@@ -1,6 +1,9 @@
 use termion::event;
 
+
+#[derive(Debug)]
 pub enum Event {
     Input(event::Key),
     Tick,
+    Quit,
 }